requirements.txt
Flask>=3.1
Flask-Mail>=0.10
psycopg2-binary>=2.9
pytz>=2024.1
